requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
cmake==3.26.4
contourpy==1.1.0
cycler==0.11.0
filelock==3.12.2
fonttools==4.40.0
idna==3.4
imageio==2.31.1
Jinja2==3.1.2
kiwisolver==1.4.4
lazy_loader==0.2
lit==16.0.6
MarkupSafe==2.1.3
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.1
numpy==1.25.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.1
Pillow==9.5.0
powerapi==2.0.4
psutil==5.9.5
pyJoules==0.5.1
pynvml==11.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
pyzmq==25.1.0
requests==2.31.0
scikit-image==0.21.0
scipy==1.10.1
setproctitle==1.3.2
six==1.16.0
sympy==1.12
tifffile==2023.4.12
torch==2.0.1
torchaudio==2.0.2
torchmetrics==0.11.4
torchvision==0.15.2
triton==2.0.0
typing_extensions==4.6.3
urllib3==2.0.3
prometheus-client==0.17.0
prometheus_api_client==0.5.3